home screen code cleanup
